Support for related lookups in detail_uri_name.
Previously, detail_uri_name could only be a direct attribute of a model rather than a related lookup.  This commit adds support for lookups that pass through related objects, allowing for URI names that depend on related attributes.

For instance, we may want URIs that look like /map/<blog post slug> for a 'blogmap' resource.  In this case it's natural to assume we would want something along the lines of detail_uri_name = 'post__slug'.